+++ gdsfactory/components/via.py
@@ -67,6 +67,6 @@ via2 = gf.partial(via, layer=LAYER.VIA2)
 if __name__ == "__main__":
 
     c = via()
-    # c.pprint
+    # c.pprint()
     print(c)
     c.show()